Premium rates fixed, now is possible add a Float number like 1.6 for 60% extra

+++ src/estaticos/AtlantaMain.kt
@@ -1567,19 +1567,19 @@ object AtlantaMain {
                             DEFECTO_DROP = RATE_DROP_NORMAL
                         }
                         "RATE_XP_PVM_ABONADOS" -> {
-                            RATE_XP_PVM_ABONADOS = valor.toInt().toFloat()
+                            RATE_XP_PVM_ABONADOS = valor.toFloat()
                             variable = "RATE_XP_PVM_ABONADOS"
                         }
                         "RATE_DROP_ABONADOS" -> {
-                            RATE_DROP_ABONADOS = valor.toInt().toFloat()
+                            RATE_DROP_ABONADOS = valor.toFloat()
                             variable = "RATE_DROP_ABONADOS"
                         }
                         "RATE_KAMAS_ABONADOS" -> {
-                            RATE_KAMAS_ABONADOS = valor.toInt().toFloat()
+                            RATE_KAMAS_ABONADOS = valor.toFloat()
                             variable = "RATE_KAMAS_ABONADOS"
                         }
                         "RATE_XP_OFICIO_ABONADOS" -> {
-                            RATE_XP_OFICIO_ABONADOS = valor.toInt().toFloat()
+                            RATE_XP_OFICIO_ABONADOS = valor.toFloat()
                             variable = "RATE_XP_OFICIO_ABONADOS"
                         }
                         "RATE_PODS" -> {
